requirements.txt
### NSFR
torch==1.8.1
lark-parser
sklearn
torchsummary
setuptools>=58.5.3
tensorboard
numpy>=1.18.5
torchvision
tqdm>=4.41.0
matplotlib>=3.2.2
opencv-python>=4.1.2
Pillow
PyYAML>=5.3.1
scipy>=1.4.1
seaborn
pandas
rtpt